+++ packages/react-native/src/icons/app/EllipsisHorizontal.tsx
@@ -1,9 +1,8 @@
 import Svg, { Circle } from "react-native-svg";
 import type { SvgProps } from "react-native-svg";
 import { Ref, forwardRef } from "react";
-const SvgEllipsisHorizontal = (props: SvgProps, ref: Ref<SVGSVGElement>) => (
+const SvgEllipsisHorizontal = (props: SvgProps, ref: Ref<Svg>) => (
   <Svg
-    xmlns="http://www.w3.org/2000/svg"
     fill="none"
     viewBox="0 0 24 24"
     className={props.className}